feat: headerProfile 분리,api 연동
오른쪽 profile 분리
api 연동

api에 유저네임 필요한데 현재 알수없어서 test로 넣어놓았음
